Removed GPUImage constructors, now in image resource manager
