+++ engine/game/private/input_bindings_visualization_cache.cpp
@@ -59,7 +59,8 @@ ResourceHandle<GPUImage> InputBindingsVisualizationCache::GetGlyph(const std::st
     }
 
     auto& imageResourceManager = _graphicsContext.Resources()->GetImageResourceManager();
-    auto image = imageResourceManager.Create(glyphImage.value(), bb::TextureFlags::COMMON_FLAGS, path);
+    auto commands = SingleTimeCommands { *_graphicsContext.GetVulkanContext() };
+    auto image = imageResourceManager.Create(glyphImage.value(), bb::TextureFlags::COMMON_FLAGS, path, &commands);
 
     _graphicsContext.UpdateBindlessSet();
     _glyphCache.emplace(path, image);
