+++ engine/resources/public/resources/texture.hpp
@@ -8,53 +8,17 @@
 #include "resource_manager.hpp"
 #include "vulkan_include.hpp"
 
-
 #include <optional>
 
 class SingleTimeCommands;
 class VulkanContext;
 
-// namespace bb
-// {
-
-// struct Texture
-// {
-//     static std::optional<Texture> fromImage(
-//         SingleTimeCommands& upload_commands,
-//         const Image2D& image,
-//         ResourceHandle<Sampler> sampler,
-//         VulkanFlags usage_flags,
-//         const char* name);
-
-//     // void destroy(VulkanContext& ctx);
-
-//     struct LayerView
-//     {
-//         VkImageView view {};
-//         std::vector<VkImageView> mips {};
-//     };
-
-//     ResourceHandle<Sampler> sampler {};
-//     VmaAllocation allocation {};
-//     VkImage image {};
-//     VkImageView full_view {};
-//     ImageFormat format {};
-//     std::vector<LayerView> layers {};
-//     ImageType type {};
-//     uint32_t width {};
-//     uint32_t height {};
-//     uint32_t depth {};
-//     uint32_t mips {};
-// };
-
-// }
-
 struct GPUImage
 {
-public:
-    GPUImage(const CPUImage& data, ResourceHandle<Sampler> textureSampler, const std::shared_ptr<VulkanContext>& context, SingleTimeCommands* const commands = nullptr);
-    GPUImage(SingleTimeCommands& upload_commands, const bb::Image2D& image, ResourceHandle<Sampler> textureSampler, bb::Flags<bb::TextureFlags> flags, std::string_view name);
+    friend class ImageResourceManager;
+    GPUImage() = default;
 
+public:
     ~GPUImage();
 
     GPUImage(GPUImage&& other) noexcept;
@@ -87,7 +51,5 @@ public:
     vk::ImageType vkType { vk::ImageType::e2D };
 
     std::string name;
-
-private:
     VulkanContext* _context;
 };
